:art: Improved the file structure
