+++ src/components/Button/AddTodo.js
@@ -1,0 +1,18 @@
+import React from "react";
+import Button from "./Button";
+import { AbsoluteBlock, RelativeBlock } from "../../styled/ButtonPosition";
+
+// For adding the Plus(+) button at the bottom
+function AddTodo() {
+  return (
+    <>
+      <RelativeBlock>
+        <AbsoluteBlock>
+            <Button />
+        </AbsoluteBlock>
+      </RelativeBlock>
+    </>
+  );
+}
+
+export default AddTodo;
