+++ src/components/Button/Button.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import {AddButton} from '../../styled/Button';
+
+
+function Button() {
+  return (
+    <AddButton>+</AddButton>
+  )
+}
+
+export default Button
