+++ src/components/InputCheckBox/InputCheckbox.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { InputCheckBox } from "../../styled/InputCheckBox";
+
+function InputCheckbox(props) {
+  const   {handleCheck,index} = props;
+  return (
+    <div>
+      <InputCheckBox
+        type="checkbox"
+        className="checkBox"
+        onChange={() => handleCheck(index)}
+      />
+    </div>
+  );
+}
+
+export default InputCheckbox;
