+++ src/constants/dummyText.js
@@ -1,0 +1,35 @@
+export const dayNames = [
+    "SUNDAY",
+    "MONDAY",
+    "TUESDAY",
+    "WEDNESDAY",
+    "THURSDAY",
+    "FRIDAY",
+    "SATURDAY",
+  ];
+
+export const monthNames = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December",
+  ];
+
+  //variable to store the list of items
+export const dummyList = [
+  "Buy new sweatshirt",
+  "Begin promotional phase",
+  "Read an article",
+  "Try not to fall asleep",
+  'Watch "Sherlock"',
+  "Begin QA for the product",
+  "Go for a walk",
+];
