--- src/css/addTodo.css
@@ -1,31 +1,0 @@
-/* Style for positioning the div to relative */
-.button-position-relative {
-    position: relative;
-    display: flex;
-    justify-content: center;
-}
-
-/* Style for positioning the div to absolute */
-.button-position-absolute {
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    top: 62px;
-}
-
-/* Style for the add button */
-.add-button {
-    border-radius: 50%;
-    height: 80px;
-    width: 80px;
-    font-size: 40px;
-    background-color: #2ceead;
-    border: none;
-    color: rgb(28, 119, 28);
-}
-
-/* Style to be applied on add button on hovering on it */
-.add-button:hover {
-    background-color: #58c9a3;
-    color: rgb(20, 85, 20);
-}
