--- src/css/navbar.css
@@ -1,25 +1,0 @@
-/* Style for the date */
-.date-style {
-    font-weight: 600;
-    font-size: 39px;
-}
-
-/* Style for the day */
-.day-style {
-    font-weight: 600;
-    font-size: 16px;
-}
-
-/* Style for the month */
-.month-style {
-    font-weight: 600;
-    position: relative;
-    top: 9px;
-    font-size: 15px;
-}
-
-/* Style for the year */
-.year-style {
-    margin-top: 2px;
-    font-size: 15px;
-}
