--- src/css/todo.css
@@ -1,8 +1,0 @@
-/* Style for main body */
-.main-body {
-    background-color: white;
-    text-align: center;
-    width: 400px;
-    margin-top: 100px;
-    height: 600px;
-}
