--- src/css/todoList.css
@@ -1,23 +1,0 @@
-/* Style for body */
-body {
-    background-color: rgb(238, 238, 238);
-}
-
-/* Style for the checkbox */
-.checkBox {
-    appearance: none;
-    border: 2px solid rgb(209, 207, 207);
-    border-radius: 50%;
-    cursor: pointer;
-    height: 28px;
-    position: relative;
-    width: 28px;
-}
-
-/* Style to be applied when the checkbox is checked */
-input[type=checkbox]:checked {
-    background-image: url(../assets/images/checkbox.ico);
-    background-repeat: no-repeat;
-    background-size: 100%;
-    border: none;
-}
